requirements.txt
aiohttp==3.5.4
async-timeout==3.0.0
atomicwrites==1.1.5
attrs==18.1.0
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
Deprecated==1.2.5
discord.py==1.2.2
fuzzywuzzy==0.17.0
idna==2.7
idna-ssl==1.1.0
more-itertools==4.2.0
multidict==4.3.1
pluggy==0.6.0
py==1.5.4
PyGithub==1.43.7
PyJWT==1.6.4
pytest==3.6.3
requests==2.19.1
six==1.11.0
typing-extensions==3.7.2
urllib3==1.23
websockets==6.0
wrapt==1.11.2
yarl==1.3.0
